DO NOT MERGE: Implement <Polling Message>.
When Device Discovery is launched or Hot-plug detection is
launched, the first step of it is to send <Polling Message>
to all remote devices. According to type of feature,
it may have different retry count for sending <Polling Message>.

As <Polling Message> to all devices should be serialized operation
it runs on io thread as single operation.

Along with this, added assertRunOnIoThread and
assertRunOnServiceThread used to make sure that all methods are
called in proper thread.

Change-Id: I2d2df0216867c188e99ba24b216ec73f3396eeae