Record properSigningKeySet values to packages.xml
PackageKeySetData records all of the defined keysets which are a subset of an
apk's signing keyset.  It also records the original superset for all of these
in an easy-to-access 'proper' signing-keyset which is consulted as the official
source for a package.  This value was not being recorded, causing
PackageManager to re-gather the certificates for a package on each boot after
a fix for a different bug.  Record it.

Bug: 16076095
Change-Id: If5a7f6d70dd4784284d7bab466dab0311aa13c28