Correct printer extra constant value.
We are passing the current printer to the custom print options activity.
Before we were using a constant specified in the java docs as it was
added past API change was possible. For this release a constant is
defined but its value was wrong and clients using the old (correct) value
could not look up the printer extra.

bug:16291822

Change-Id: Iaff6728865690caa8aaaa49a5911009d9f099148